#338 사용자 프로필 이동 구현 및 상세 화면 API 연동

+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/response/comment/ReplyResponse.kt
@@ -13,7 +13,7 @@ data class ReplyResponse(
     val isLikedByMe: Boolean,
     val likeCount: Int,
     val member: MemberResponse,
-    val replyId: Int,
+    val replyId: Long,
 )
 
 fun ReplyResponse.toDomain() =
